+++ lib/src/arbitrary/arbitrary.dart
@@ -15,6 +15,7 @@ export 'top.dart'
         constantFrom,
         dateTime,
         deck,
+        duration,
         float,
         frequency,
         integer,
@@ -26,4 +27,5 @@ export 'top.dart'
         recursive,
         runes,
         set,
-        string;
+        string,
+        uri;